Sleepwalker does produce None on 204: default it.
This goes back to an older idea of the default request/response
being None (i.e. a 204, as a 202 must have a valid JSON document as
a body, and while the empty string is a valid piece of JSON data,
it is not a valid document).

It turns out sleepwalker does produce this value correctly, so
go ahead and stop trying to accomodate the requests library's behavior
of returning an empty string.