Don't intercept 'links' and 'relations' as keys.
For presumably historical reasons, we were intercepting the property
names 'links' and 'relations' and mapping them to self.links and
self.relations.  There is no way that people will always avoid
those strings as property names, and even if the property would
override this behavior, having this inconsistency is too confusing.