Implement support for servicedef id, $schema, and proper $ref
This version of reschema requires this $schema in the service
definition:

  $schema: "http://support.riverbed.com/apis/service_def/2.1"
  id: "http://support.riverbed.com/apis/test/1.0"
  provider: "riverbed"
  name: "test"
  version: "1.0"

The id must end in with {name}/{version} according to the
name and version fields.

All $ref and relation resource references must be of the form:

  $ref: "#/resources/book"
  $ref: "#/types/address"

Currently only local references to other resources and types
in the same service definition are supported.