ServiceDefCache a normal object (not singleton), expand_id/find normal methods
Based on discussions with Henry, I agree that it's better long term to
drop the "singleton" idea of ServiceDefCache.  In particular, it allows
for subclassing and more direct maintenance.  Note that this means that
a cache must be attached to a ServiceDef in order to follow references
via find to another service.

Related expand_id and find off of ServiceDef are now just normal
methods.  The only hypothetical case was "bootstrapping" where you have
a resource id (service def id + resource fragment), but can't convince
myself this is needed.  If a use is found we can resurrect later as it's
own named class method.