Support for $merge operation
This commit adds support for $merge as follows:

  {
    '$merge': {
      'source': <schema>,
      'with': <object>
    }
  }

The <schema> can be either an inline schema or a $ref.  <object>
is any partical schema that is used to update <schema> according
to http://tools.ietf.org/html/draft-ietf-appsawg-json-merge-patch-02.

As part of this, code has changed to restrict $ref to be compliant
with http://tools.ietf.org/html/draft-pbryan-zyp-json-ref-03, which
states that a $ref may have no other members:

Valid:

   { '$ref': <jsonpointer> }

Invalid:

   { '$ref': <jsonpointer>, 'description': 'My reference!' }

The above invalid combination can now be achieved using $merge:

Valid:
   {
     '$merge': {
       'source': { '$ref': <jsonpointer> },
       'with': { 'description': 'My reference!' }
     }
   }