Bug 165906 - Fix reschema-doc handling of $ref
This includes some changes to the command line arguments as well
as where the output files are generated:

Basic:

$ reschema-doc -f <inputfile> -o <outdir>

It will write the output HTML to

	<outdir>/<name>/<version>/service.html

Any references to types/resources in other schemas will generate links
based on the same above format.  The $refs are not checked as part of
the HTML generation, but instead are (blindly) assumed to be valid.
That means you can generate documentation for the files independently,
but then pull all the HTML docs together in the same tree based on
above.

The references are all relative, so you can relocate the tree without a
problem.

This commit also moves the basics of reschema-doc to reschema/doc.py
to make testing of the code the same.