"Catalog" => "bookstore", ".yml" => ".yaml"
Because repeated experience shows that lots of people think that
"catalog" refers to the schema generically, rather than being
specific to the collection-of-books example.  So call it a bookstore.

Also, "*.yaml" is preferred over "*.yml"

According to the YAML FAQ.  Also, we use *.yaml everywhere
internally so having the examples use *.yml is confusing.