Update and clarify path template/vars behavior.
While vars with a schema rather than a relative json pointer are not
yet implemented in sleepwalker, we are also not yet open sourcing
reschema and sleepwalker so we can just consider that an internal bug
and go ahead and release the full specification.

Therefore mark the params: field deprecated (it's no longer needed) and
explain the restricitons on uritemplates and schema more clearly.
The uritemplate rules already cover how to handle resolving array and
object values when the correct template syntax is used.