Allow pre-processing as a class method.
The resolve_defs method (plural) is just a wrapper around the resolve_def
(singluar) classmethod, and therefore can also be a classmethod.  Anything
calling it as an instance method will still get the same behavior.

This means that preprocess can be a classmethod, but leave behind the original
instancemethod with its original name as the behavior is slightly different
and is in active use.

This allows for preprocessing data before handing it to Schema.parse()
that does ont require its own parser instance for anything else.

Also add unit tests, flake8 fixes.

ReviewId: 116476