Factor out a class to deal with id & name.
There was a lot of duplication in id, name and repr handling.  And where
there was not duplication, the differeneces were a bit confusing.
I'm not sure the different name generation options are much more clear
this way, but I feel that it's easier to see the overall shape of things
when grouped this way.

Also, use super() properly and consistently.

ReviewId: 116212